updated requirements

+++ etl/sql/create_fact_rsi.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS fact_rsi (
     timeframe_id INTEGER NOT NULL,
     period INTEGER NOT NULL,
     calc_version TEXT NOT NULL,
-    value NUMERIC(10, 4),
+    value FLOAT,
     calculated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     PRIMARY KEY (time, currency_id, timeframe_id, period, calc_version)
 );
